Team_Birch/coursework_two/pyproject.toml
[tool.poetry]
name = "coursework-two"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
pandas = ">=2.2.3,<3.0.0"
sqlalchemy = ">=2.0.40,<3.0.0"
pyyaml = ">=6.0.2,<7.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
pypdf = ">=4.0.0,<5.0.0"
fuzzywuzzy = "^0.18.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.5"
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
bandit = "^1.8.3"
safety = "^3.3.1"
pre-commit = "^4.2.0"
sphinx-autoapi = "^3.6.0"
sphinx-rtd-theme = "^3.0.2"
requests = "^2.32.3"
python-dotenv = "^1.1.0"

[[tool.poetry.packages]]
include = "modules"
from = "src"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = [
  "integration: mark a test as integration"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
